got disable auto accept button working
