+++ app/webpacker/components/RegistrationsV2/api/registration/patch/update_registration.js
@@ -29,15 +29,3 @@ export async function bulkUpdateRegistrations(
   });
   return data;
 }
-
-export async function bulkAutoAcceptRegistrations(competitionId) {
-  const route = bulkAutoAcceptRegistrationsUrl(competitionId);
-  const { data } = await fetchWithJWTToken(route, {
-    method: 'GET',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  });
-  return data;
-}
-
