+++ app/webpacker/lib/requests/routes.js.erb
@@ -282,4 +282,6 @@ export const bulkUpdateRegistrationUrl = `<%= CGI.unescape(Rails.application.rou
 
 export const bulkAutoAcceptRegistrationsUrl = (competitionId) => `<%= CGI.unescape(Rails.application.routes.url_helpers.api_v1_registrations_bulk_auto_accept_path(competition_id: "${competitionId}")) %>`;
 
+export const disableAutoAcceptUrl = (competitionId) => `<%= CGI.unescape(Rails.application.routes.url_helpers.api_v0_competition_disable_auto_accept_path(competition_id: "${competitionId}")) %>`;
+
 export const paymentTicketUrl = (competitionId, donationIso) => `<%= CGI.unescape(Rails.application.routes.url_helpers.api_v1_registrations_payment_ticket_path(competition_id: "${competitionId}", donation_iso: "${donationIso}")) %>`;
