+++ config/locales/en.yml
@@ -1850,8 +1850,8 @@ en:
         next_step: "Continue to next Step"
       update:
         bulk_auto_accept: "Bulk Auto Accept"
-        enable_bulk_auto_accept: "Enable Auto Accept"
-        disable_bulk_auto_accept: "Disable Auto Accept"
+        enable_auto_accept: "Enable Auto Accept"
+        disable_auto_accept: "Disable Auto Accept"
         being_updated: "Registration is being updated..."
         being_deleted: "Registration is being deleted..."
         update_confirm: "Are you sure you want to update your registration?"
